clean up code, move functions out of index.js
